Added support for selecting eeprom with verified correct data.
Added a data pointer to the eeprom structure that is usually null but set to the address
of data loaded onto the heap when read_write_eeprom is called to read.
Modified verify_eeprom to clear that data if the sha does not match.
Added to the initialize function to find a good eeprom, then reset all bad eeproms.
While it scans for a good eeprom, the data for that good eeprom gets loaded and only
that data remains. When resetting bad eeproms, it uses that good_eeprom's same data
address.