Fixed issue where complete path was not obtained.
Code previously used fgetc to see if the next character was a comment line and skipped the line if it was
but that also meant that it dropped the first character from the path when it was not.
Modified to skip the line when fscanf does not get 3 fields or if the path field starts with a '#'.