Fixed find_good_eeprom.
Previously had a segfault where scanning the max_wc_eeprom list would go past the end and end up dereferencing a
pointer to some random address. Fixed by adding an additional check to make sure that the loop hasn't passed the
end of the array.
Also changed the way the first item is loaded into max_wc_eeprom. Instead of just immediately putting the
first item in there, it checks that the first item is null, then loads it and continues.