Fixed clear_after_null and its use.
Clear after null previously returned the length of the string it cleared because of bad logic in the if statment that set r.
Fixed by ony setting r if r was previously not set. This also meant that the if determining if the byte should be cleared
needed to also be changed to trigger if r is greater than or equal to 0.

Also fixed various places where the return from clear_after_null is used. Mostly needed to change if (null_found) to
if (null_found >= 0).

Finally, fixed the return from read_write_eeprom to be the actual correct length of string read. Previously counted
null byte at end as part of the string length.