Fixed a bug where eeprom contents were written to stdout.
Forgot to add open_eeproms() / close_eeproms() to the eeprom_manager_set_set() and eeprom_mangager_get_value() api functions.
This caused them to have fds of 0 which caused their contents to be written to stdout.
Also added a check for device->fd being 0 to the read_write_all function to return -1 with errno set to EINVAL.