Implemented proper json_root object caching.
Modified update_eeprom_data to return 1 when data has changed.
Updated the eeprom_manager_read_value and eeprom_manager_set_value functions to only free the json_root object
if its initialization fails or the return from update_eeprom_data indicates that the data has changed.
Then it checks if the json_root object is NULL and loads it based on that condition.
Had to add a quick check and free to the eeprom_manager_cleanup function.