Fixed clear command with no good devices.
Previously didn't have logic to check if the user was trying to clear the eeprom devices when init returned the
EEPROM_MANAGER_ERROR_NO_GOOD_DEVICES_FOUND error. Added logic to not disply the error and exit when the user
is trying to clear them. Added an additional clause to the catch-all case to not catch EEPROM_MANAGER_ERROR_NO_GOOD_DEVICES_FOUND
errors since they were either already caught or ignored.